fix: sip10 tokens ledger send form, closes #4491
