+++ src/app/features/asset-list/asset-list.tsx
@@ -6,18 +6,14 @@ import { Stack } from 'leather-styles/jsx';
 import { LEDGER_BITCOIN_ENABLED } from '@shared/environment';
 
 import { useBtcAssetBalance } from '@app/common/hooks/balance/btc/use-btc-balance';
-import { useStxBalance } from '@app/common/hooks/balance/stx/use-stx-balance';
 import { useWalletType } from '@app/common/use-wallet-type';
 import { BitcoinContractEntryPoint } from '@app/components/bitcoin-contract-entry-point/bitcoin-contract-entry-point';
 import { Brc20TokensLoader } from '@app/components/brc20-tokens-loader';
 import { CryptoCurrencyAssetItem } from '@app/components/crypto-assets/crypto-currency-asset/crypto-currency-asset-item';
-import { StxAvatar } from '@app/components/crypto-assets/stacks/components/stx-avatar';
 import { BtcIcon } from '@app/components/icons/btc-icon';
 import { CurrentStacksAccountLoader } from '@app/components/stacks-account-loader';
 import { useHasBitcoinLedgerKeychain } from '@app/store/accounts/blockchain/bitcoin/bitcoin.ledger';
 import { useCurrentAccountNativeSegwitAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
-import { useCurrentStacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
-import { useHasStacksLedgerKeychain } from '@app/store/accounts/blockchain/stacks/stacks.hooks';
 import { useCurrentNetwork } from '@app/store/networks/networks.selectors';
 
 import { Collectibles } from '../collectibles/collectibles';
@@ -25,16 +21,14 @@ import { PendingBrc20TransferList } from '../pending-brc-20-transfers/pending-br
 import { BitcoinFungibleTokenAssetList } from './components/bitcoin-fungible-tokens-asset-list';
 import { ConnectLedgerAssetBtn } from './components/connect-ledger-asset-button';
 import { StacksBalanceItem } from './components/stacks-balance-item';
+import { StacksLedgerAssetsList } from './components/stacks-ledger-assets';
 
 export function AssetsList() {
-  const hasStacksKeys = useHasStacksLedgerKeychain();
   const hasBitcoinLedgerKeys = useHasBitcoinLedgerKeychain();
   const btcAddress = useCurrentAccountNativeSegwitAddressIndexZero();
   const network = useCurrentNetwork();
-  const currentAccount = useCurrentStacksAccount();
 
   const { btcAvailableAssetBalance, btcAvailableUsdBalance } = useBtcAssetBalance(btcAddress);
-  const { stxEffectiveBalance, stxEffectiveUsdBalance } = useStxBalance();
 
   const { whenWallet } = useWalletType();
 
@@ -74,15 +68,7 @@ export function AssetsList() {
             {account => <StacksBalanceItem account={account} />}
           </CurrentStacksAccountLoader>
         ),
-        ledger: (
-          <CryptoCurrencyAssetItem
-            assetBalance={stxEffectiveBalance}
-            usdBalance={stxEffectiveUsdBalance}
-            icon={<StxAvatar />}
-            address={currentAccount?.address || ''}
-            rightElement={hasStacksKeys ? undefined : <ConnectLedgerAssetBtn chain="stacks" />}
-          />
-        ),
+        ledger: <StacksLedgerAssetsList />,
       })}
 
       {whenWallet({
